requirements.txt
requests
beautifulsoup4
pymongo[tls,srv]
python-dotenv